+++ src/Rendering/Font/TextFlow.h
@@ -66,7 +66,7 @@ namespace eyegui
         {
             std::shared_ptr<std::vector<glm::vec3> > spVertices;
             std::shared_ptr<std::vector<glm::vec2> > spTextureCoordinates;
-            float width;
+            int width;
         };
 
         // Calculate mesh (in pixel coordinates)
